Added Pedestal menu, hammer only mines blocks in it's set, shows a block highlight, and the breaking animation

+++ src/main/java/net/j40climb/florafauna/event/Events.java
@@ -1,6 +1,7 @@
 package net.j40climb.florafauna.event;
 
 import net.j40climb.florafauna.FloraFauna;
+import net.j40climb.florafauna.common.items.interfaces.DiggerTool;
 import net.j40climb.florafauna.component.DataComponentTypes;
 import net.j40climb.florafauna.item.ModItems;
 import net.j40climb.florafauna.item.custom.HammerItem;
@@ -39,19 +40,24 @@ public class Events {
 
             int range = Objects.requireNonNull(mainHandItem.get(DataComponentTypes.HIT_RANGE)).range();
 
-            for(BlockPos pos : HammerItem.getBlocksToBeDestroyed(range, initialBlockPos, serverPlayer)) {
-                // Don't mine the position that was just mined, or a different type of block
-                if(pos == initialBlockPos || !hammer.isCorrectToolForDrops(mainHandItem, event.getLevel().getBlockState(pos))) {
-                    continue;
-                }
+            // Only do a hammer mine if the block being mined is the correct tool
+            if (hammer.isCorrectToolForDrops(mainHandItem, event.getLevel().getBlockState(initialBlockPos))) {
+                for(BlockPos pos : DiggerTool.getBlocksToBeBroken(range, initialBlockPos, serverPlayer)) {
+                    // Don't mine the position that was just mined, or a different type of block
+                    if(pos == initialBlockPos || !hammer.isCorrectToolForDrops(mainHandItem, event.getLevel().getBlockState(pos))) {
+                        continue;
+                    }
 
-                HARVESTED_BLOCKS.add(pos);
-                // Call destroy which runs this event again so we need the fast return above in if(HARVESTED_BLOCKS.contains(initialBlockPos))
-                // If not, it would keep deleting and not move through the set, creating an infinite loop
-                // https://courses.kaupenjoe.net/courses/modding-by-kaupenjoe-neoforge-modding-for-minecraft-1-21-x/lectures/55341862 at 10min in
-                serverPlayer.gameMode.destroyBlock(pos);
-                HARVESTED_BLOCKS.remove(pos);
+                    HARVESTED_BLOCKS.add(pos);
+                    // Call destroy which runs this event again so we need the fast return above in if(HARVESTED_BLOCKS.contains(initialBlockPos))
+                    // If not, it would keep deleting and not move through the set, creating an infinite loop
+                    // https://courses.kaupenjoe.net/courses/modding-by-kaupenjoe-neoforge-modding-for-minecraft-1-21-x/lectures/55341862 at 10min in
+                    serverPlayer.gameMode.destroyBlock(pos);
+                    HARVESTED_BLOCKS.remove(pos);
+                }
             }
+
+
         }
     }
